// vendor_app/src/api/models/ItemCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Vendor Management API
 * API for managing vendor items
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ItemCreate
 */
export interface ItemCreate {
    /**
     * 
     * @type {string}
     * @memberof ItemCreate
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ItemCreate
     */
    description: string;
    /**
     * 
     * @type {number}
     * @memberof ItemCreate
     */
    price: number;
}

/**
 * Check if a given object implements the ItemCreate interface.
 */
export function instanceOfItemCreate(value: object): value is ItemCreate {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('price' in value) || value['price'] === undefined) return false;
    return true;
}

export function ItemCreateFromJSON(json: any): ItemCreate {
    return ItemCreateFromJSONTyped(json, false);
}

export function ItemCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ItemCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'description': json['description'],
        'price': json['price'],
    };
}

export function ItemCreateToJSON(json: any): ItemCreate {
    return ItemCreateToJSONTyped(json, false);
}

export function ItemCreateToJSONTyped(value?: ItemCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'description': value['description'],
        'price': value['price'],
    };
}